+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Version 1.2.1
 
-Released 2021-02-XX
+Released 2021-02-12
 
+- Errors are more verbose, suggest fixes based off of error handling and have
+a more human syntax
 - Bugfix: File Handling:
   - Previously when the user had read permission to a file, but no ownership,
 the file was copied, but an error was thrown saying that access was denied.